pyproject.toml
[tool.poetry]
name = "reddit2spotify"
version = "0.1.0"
description = "A script to create spotify playlists based on reddit recommendations"
authors = ["Rishi Banerjee <rishieric91@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
praw = "^7.7.1"
spotipy = "^2.24.0"
python-dotenv = "^1.0.1"
halo = "^0.0.31"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
